116.dockter.1 (#117)
* #116 Migrate from int to int64 where needed

* #116 Migrate from int to int64 where needed

* #116 Migrate from int to int64 where needed

* #116 Improve examples

* #116 Improve test names

+++ g2config/g2config_examples_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package g2config
 
 import (
@@ -163,7 +165,7 @@ func ExampleG2config_Init() {
 	if err != nil {
 		fmt.Println(err)
 	}
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	err = g2config.Init(ctx, moduleName, iniParams, verboseLogging)
 	if err != nil {
 		fmt.Println(err)
